DAPI Injector: dll_path is char[] so sizeof works
As originally written, the type for `dll_path` is deduced to be char*, and (as a pointer type) sizeof returns 4 for 32bit systems.

In order to get the desired behavior, `dll_path` must be explicitly an array. That way, sizeof returns the actual string length, which is 9 in this case.

Since 4 != 9 before, not enough space would be allocated for the string so the LoadLibrary would fail.